+++ lib/src/dart/imports.dart
@@ -3,16 +3,12 @@ import 'package:reduct/reduct.dart';
 String addImports(List<Column> columns) {
   final buffer = StringBuffer();
 
-  bool hasDecimal = false;
   bool hasDate = false;
   bool hasTimestamptz = false;
   for (var column in columns) {
     switch (column.type) {
       case ColumnTypeDuckDB.date:
         hasDate = true;
-        break;
-      case ColumnTypeDuckDB.decimal:
-        hasDecimal = true;
         break;
       case ColumnTypeDuckDB.timestamptz:
         hasTimestamptz = true;
@@ -22,12 +18,11 @@ String addImports(List<Column> columns) {
     }
   }
 
+  buffer.writeln("import 'dart:convert';");
+  buffer.writeln();
   buffer.writeln("import 'package:http/http.dart' as http;");
   if (hasDate) {
     buffer.writeln("import 'package:date/date.dart';");
-  }
-  if (hasDecimal) {
-    buffer.writeln("import 'package:decimal/decimal.dart';");
   }
   if (hasTimestamptz) {
     buffer.writeln("import 'package:timezone/timezone.dart';");
@@ -35,4 +30,3 @@ String addImports(List<Column> columns) {
   buffer.writeln();
   return buffer.toString();
 }
-